extract decompression-stream into its own file
